feat: implement JPA-based company and job offer management
- add JPA/H2 setup and test application properties
- add/adjust Company and JobOffer models, repositories, and services
- expose JobOffer CRUD endpoints and remove HelloController
- add global exception handling with structured error responses
- add integration and service tests for Company and JobOffer flows
- add Qodana configuration

+++ src/main/java/com/cgorin/jobtracker/exception/CompanyAlreadyExistsException.java
@@ -1,7 +1,7 @@
 package com.cgorin.jobtracker.exception;
 
 public class CompanyAlreadyExistsException extends RuntimeException {
-    public CompanyAlreadyExistsException(int id) {
+    public CompanyAlreadyExistsException(Long id) {
         super("Company"  + id + " already exists");
     }
 }
