+++ src/main/java/com/cgorin/jobtracker/exception/CompanyNotFoundException.java
@@ -1,7 +1,7 @@
 package com.cgorin.jobtracker.exception;
 
 public class CompanyNotFoundException extends RuntimeException {
-    public CompanyNotFoundException(int id) {
+    public CompanyNotFoundException(Long id) {
         super("Company with id " + id + " not found");
     }
 }
