+++ src/main/java/com/cgorin/jobtracker/exception/JobOfferNotFoundException.java
@@ -1,7 +1,7 @@
 package com.cgorin.jobtracker.exception;
 
 public class JobOfferNotFoundException extends RuntimeException {
-    public JobOfferNotFoundException(int id) {
+    public JobOfferNotFoundException(Long id) {
         super("Job offer with id " + id + " not found");
     }
 }
